Use getFileHash in installAddon HOC
